\name{inversions}
\alias{inversions}
\title{
Count the number of total inversions of a permutation
}
\description{
inversions() counts the number of inversions of a permutation.
}
\usage{
inversions(x)
}
\arguments{
  \item{x}{ a permutation of 1, 2, ..., n.
}
}
\details{
Input must be a list of numbers.
}
\value{
The total number of inversions in a list.
}


\seealso{
 \code{\link{inversionv}}
}
\examples{
inversions(c(3,1,2))


}

\keyword{ Permutation}
